# openapi.yaml
openapi: 3.0.0
info:
  title: Learning Management System (LMS) API
  description: API для системы управления обучением в колледжах
  version: 1.0.0
servers:
  - url: http://localhost:8091
    description: Development server
  - url: https://api.college.ru
    description: Production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
...
paths:
  /login:
    post:
      tags: [Authentication]
      summary: User login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
              required: [email, password]
      responses:
        '200':
          description: Successful login
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          description: Invalid request
        '401':
          description: Invalid credentials
        '500':
          description: Internal server error
  /refresh:
    post:
      tags: [Authentication]
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refresh_token:
                  type: string
              required: [refresh_token]
      responses:
        '200':
          description: New access token generated
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
        '400':
          description: Invalid request
        '401':
          description: Invalid or expired refresh token
        '500':
          description: Internal server error
...